python_projects_grocery_webapp/requirements.txt
flask==2.0.1
flask-cors==3.0.10
mysql-connector-python==8.0.26
python-dotenv==0.19.0
werkzeug==2.0.1
gunicorn==20.1.0
